+++ Code/Engine/Renderer/VertexBuffer.cpp
@@ -11,7 +11,7 @@ VertexBuffer::VertexBuffer(RHIDevice* owner, const std::vector<Vertex3D>& buffer
     D3D11_BUFFER_DESC vb_desc;
     memset(&vb_desc, 0, sizeof(vb_desc));
 
-    vb_desc.ByteWidth = sizeof(Vertex3D) * buffer.size();
+    vb_desc.ByteWidth = sizeof(Vertex3D) * static_cast<unsigned int>(buffer.size());
     vb_desc.BindFlags = BufferBindUsageToD3DBindFlags(bindUsage);
     vb_desc.Usage = BufferUsageToD3DUsage(usage);
     vb_desc.StructureByteStride = sizeof(Vertex3D);
